+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in each version of the haproxy cookbook.
 
 ## [unreleased]
 
+* Added Chefspec Matchers for the resources defined in this cookbook.
+
 ## [v4.2.0] (04-05-2017)
 
 - Added in `acl` resource, usage: `test/fixtures/cookbooks/test/recipes/config_acl.rb`