+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file is used to list changes made in each version of the haproxy cookbook.
 
 * Added Chefspec Matchers for the resources defined in this cookbook.
 * Added `mode` property to `backend` and `frontend` resources.
+* Added `maxconn` to `global` resource
 
 ## [v4.2.0] (04-05-2017)