+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This file is used to list changes made in each version of the haproxy cookbook.
 
 ## [unreleased]
 
+
+## [v4.3.0] (05-31-2017)
+
 * Added Chefspec Matchers for the resources defined in this cookbook.
 * Added `mode` property to `backend` and `frontend` resources.
 * Added `maxconn` to `global` resource
@@ -267,7 +270,7 @@ This file is used to list changes made in each version of the haproxy cookbook.
 
 - Use `node.chef_environment` instead of `node['app_environment']`
 
-[unreleased]: https://github.com/sous-chefs/haproxy/compare/v4.2.0...HEAD
+[unreleased]: https://github.com/sous-chefs/haproxy/compare/v4.3.0...HEAD
 [v3.0.0]: https://github.com/sous-chefs/haproxy/compare/v2.0.2...v3.0.0
 [v3.0.1]: https://github.com/sous-chefs/haproxy/compare/v3.0.0...v3.0.1
 [v3.0.2]: https://github.com/sous-chefs/haproxy/compare/v3.0.1...v3.0.2
@@ -278,3 +281,4 @@ This file is used to list changes made in each version of the haproxy cookbook.
 [v4.0.2]: https://github.com/sous-chefs/haproxy/compare/v4.0.1...v4.0.2
 [v4.1.0]: https://github.com/sous-chefs/haproxy/compare/v4.0.2...v4.1.0
 [v4.2.0]: https://github.com/sous-chefs/haproxy/compare/v4.1.0...v4.2.0
+[v4.3.0]: https://github.com/sous-chefs/haproxy/compare/v4.2.0...v4.3.0